Predictions for glmELE Objects

Obtains predictions and optionally standard errors from a fitted glmELE object. This method handles new data, factor level consistency, offsets, and confidence intervals.

Usage

# S3 method for class 'glmELE'
predict(
  object,
  newdata = NULL,
  weight.matrix = NULL,
  type = c("link", "response"),
  se.fit = FALSE,
  interval = c("none", "confidence"),
  level = 0.95,
  na.action = stats::na.pass,
  ...
)

Arguments

object

An object of class "glmELE".

newdata

An optional data frame in which to look for variables with which to predict. If omitted, the fitted linear predictors from the object are used.

weight.matrix

Character string specifying the weighting method to use (e.g., "ratio", "LL", "BLUE"). Defaults to the first method found in the object.

type

The type of prediction required. The default is on the scale of the linear predictors; the alternative "response" is on the scale of the response variable.

se.fit

Logical; if TRUE, standard errors are returned.

interval

Type of interval calculation. Can be "none" or "confidence".

level

The confidence level required (default 0.95).

na.action

A function determining what should be done with missing values in newdata. The default is to predict NA.

...

Additional arguments passed to methods.

Value

If se.fit = FALSE and interval = "none", a vector of predictions. Otherwise, a list containing:

fit

Predictions (or a matrix with columns fit, lwr, upr if intervals are requested).

se.fit

Estimated standard errors (if requested).

residual.scale

The dispersion parameter used.

Examples

data(brfss, package = "postlink")

adj_object <- adjELE(linked.data = brfss,
                     m.rate = unique(brfss$m.rate),
                     blocks = imonth,
                     weight.matrix = "BLUE")

fit <- plglm(Weight ~ Height + Physhlth + Menthlth + Exerany,
             family = "gaussian", adjustment = adj_object)

predict(fit)
#>        1        2        3        4        5        6        7        8 
#> 109.4223 171.9161 168.3558 166.6456 164.3712 164.9592 164.1647 167.9586 
#>        9       10       11       12       13       14       15       16 
#> 169.7349 161.7561 170.2268 169.8175 182.2016 176.1284 171.9726 166.7948 
#>       17       18       19       20       21       23       24       25 
#> 163.8527 172.3696 221.7342 170.6575 178.3050 221.7342 173.7141 166.8548 
#>       26       27       28       29       30       31       32       33 
#> 164.1082 170.5675 166.8548 167.0483 164.2935 173.0096 166.4603 167.6363 
#>       34       35       36       37       38       39       40       41 
#> 163.5767 174.7217 112.6970 170.4620 227.3609 164.1082 163.9528 164.6962 
#>       42       43       44       45       46       47       48       49 
#> 221.1461 222.9102 164.1082 174.1829 164.1082 162.3441 174.4988 167.6363 
#>       50       51       52       53       54       55       56       57 
#> 167.0346 162.3441 173.5457 166.7826 164.1082 163.5202 166.4603 230.4309 
#>       58       59       60       61       62       63       64       65 
#> 220.5581 164.6962 180.5791 164.1082 168.9514 163.5202 220.5581 220.5581 
#>       66       67       68       69       70       71       73       74 
#> 164.1082 167.6363 169.2477 170.9110 170.2601 165.2842 179.1022 167.3829 
#>       75       77       78       79       80       81       82       83 
#> 177.5675 185.9661 162.9322 162.3441 164.6962 168.5589 165.7196 172.5223 
#>       84       85       86       87       88       89       90       91 
#> 161.7561 164.6962 173.6801 164.6962 166.4603 169.3776 222.0564 168.2243 
#>       92       93       94       95       96       97       98       99 
#> 164.2397 162.9322 169.7349 168.8689 165.8723 162.3441 168.5589 167.0483 
#>      100      101      102      103      104      105      106      107 
#> 173.4937 162.3441 173.0538 167.0483 174.1856 169.9952 165.8723 220.5581 
#>      108      109      110      111      112      113      114      115 
#> 168.5589 174.0817 177.7215 162.9322 163.5202 167.0483 169.7349 170.9110 
#>      116      117      118      119      120      121      122      123 
#> 166.6688 168.2216 170.3230 167.6363 165.2842 174.2994 119.4342 181.8377 
#>      124      125      126      127      128      129      130      131 
#> 162.9322 166.4603 163.5202 165.4928 222.3917 171.4990 224.2123 169.9202 
#>      132      133      134      135      136      137      138      139 
#> 163.8990 166.3363 164.0305 223.2863 165.2842 166.1353 167.0483 185.0298 
#>      140      141      142      143      145      146      147      148 
#> 164.9048 221.1461 167.0483 164.6962 225.6534 169.2164 163.5202 167.6363 
#>      149      150      151      152      153      154      155      156 
#> 168.6284 164.1082 172.0966 164.6962 165.2842 165.5473 179.4856 162.9322 
#>      157      158      160      161      163      164      165      166 
#> 222.9047 179.9911 172.0870 176.3331 243.0729 221.1461 166.4603 168.2339 
#>      167      168      169      170      171      172      173      174 
#> 167.1736 174.1856 164.1082 165.2842 171.9651 171.4243 179.9865 221.1461 
#>      175      176      177      178      180      181      182      183 
#> 169.6648 165.4157 167.6363 168.2243 166.2068 165.3538 164.6962 174.1856 
#>      184      185      186      187      188      189      190      191 
#> 168.3830 167.4959 123.5412 233.3099 168.5589 165.2842 165.8723 175.3617 
#>      192      193      194      195      196      197      198      199 
#> 165.8723 168.7850 170.6575 165.8723 162.9322 167.3829 222.3222 171.1416 
#>      200      201      202      203      204      205      206      207 
#> 110.1984 164.1082 166.4603 165.8723 162.6072 165.2842 165.8723 170.3230 
#>      208      209      210      211      212      213      214      215 
#> 164.9788 164.6962 168.2243 165.8723 163.2647 221.1461 165.0185 223.6651 
#>      216      217      218      219      220      221      222      223 
#> 164.3712 172.0870 164.1082 166.2668 167.0483 169.6830 165.2842 169.1469 
#>      224      225      226      227      228      229      230      231 
#> 221.7342 163.9147 165.2842 165.2842 161.7561 178.6715 162.9322 108.8343 
#>      232      233      234      235      236      237      238      239 
#> 172.0870 170.3230 221.7342 170.4055 166.8521 187.0024 221.5249 164.6962 
#>      240      241      242      243      244      245      246      247 
#> 176.1284 167.6363 186.2884 173.8283 174.0817 170.7018 221.7342 164.7657 
#>      248      249      250      251      252      253      254      255 
#> 181.6375 170.4360 164.1055 175.8184 162.9322 166.4603 165.2842 166.2511 
#>      256      257      258      259      260      261      262      263 
#> 162.8702 162.9322 165.2842 180.9962 163.3832 162.0784 166.3983 174.7737 
#>      264      265      266      267      268      269      270      271 
#> 175.8353 168.4662 175.6174 167.0483 167.1798 221.0842 166.5141 231.2236 
#>      272      273      274      275      276      277      278      279 
#> 171.2333 165.8723 177.6139 176.9250 167.6363 221.1461 223.0417 178.1308 
#>      280      281      282      283      284      285      286      287 
#> 166.7826 284.0738 164.1082 173.9986 165.8723 168.2243 226.5682 221.1461 
#>      288      289      290      291      292      293      294      295 
#> 164.1082 220.5581 170.9110 169.1292 167.6363 221.1461 164.1082 164.3037 
#>      296      297      298      299      300      301      302      303 
#> 112.1090 224.2253 167.6363 167.6363 165.8723 164.6962 221.2027 172.0870 
#>      304      305      306      307      308      309      310      311 
#> 178.8976 167.3829 165.8723 163.5202 225.6579 166.7948 164.4305 164.1082 
#>      312      313      314      315      316      317      318      319 
#> 167.8993 238.6222 167.6363 165.2842 167.9144 164.9592 168.3034 230.6356 
#>      320      321      322      323      324      325      326      327 
#> 164.1082 178.6216 165.8723 167.0483 120.5815 168.6031 166.4603 114.1360 
#>      328      329      330      331      332      333      334      335 
#> 175.8127 167.6363 118.3237 170.1295 167.3113 164.0250 221.7342 165.8723 
#>      336      337      338      339      340      341      342      343 
#> 166.4603 166.4603 165.2842 230.2263 168.6031 221.1461 166.4603 166.7233 
#>      344      345      346      347      348      349      350      351 
#> 176.5455 179.4031 163.5202 225.4033 171.4990 169.2203 165.3538 169.4692 
#>      352      353      356      357      358      359      360      361 
#> 220.8804 172.0870 168.2243 164.7527 175.3617 177.1257 168.2243 178.2270 
#>      362      363      364      365      366      367      368      369 
#> 168.2809 170.3018 165.0185 166.4603 162.3441 165.7196 167.5219 229.4595 
#>      370      371      372      373      374      375      376      377 
#> 170.1703 164.1082 164.6962 167.6363 167.6363 167.3829 230.6356 168.5070 
#>      378      379      380      381      382      384      385      386 
#> 170.3230 167.5524 164.7527 167.0483 171.4990 220.8211 224.1373 221.7342 
#>      387      388      389      390      391      392      393      394 
#> 170.6575 172.0870 170.9110 168.8370 232.4730 165.8723 109.1566 167.6363 
#>      395      399      400      401      402      403      404      405 
#> 164.1082 174.3755 164.1647 171.4990 225.0088 177.1257 168.2243 164.1082 
#>      406      407      408      409      410      411      412      413 
#> 165.2842 220.5581 164.1082 163.5202 225.5968 167.9709 171.8372 226.1359 
#>      414      415      416      417      418      419      420      421 
#> 162.9322 165.2842 172.9030 168.5254 162.9322 175.8146 165.5206 175.1146 
#>      422      423      424      425      426      427      428      429 
#> 164.1082 170.9110 168.4329 168.1819 163.5202 168.8819 167.0483 171.4990 
#>      430      431      432      433      434      435      436      437 
#> 165.2842 163.5202 176.5455 172.9365 176.5455 169.7349 165.2842 166.4603 
#>      438      439      440      441      442      443      444      445 
#> 164.6962 163.5202 170.3230 172.1481 163.5202 221.1461 168.2243 220.5581 
#>      447      448      449      450      451      452      453      454 
#> 167.6363 163.5202 165.2842 183.0007 166.3048 222.1695 161.7561 220.8211 
#>      455      456      457      458      459      460      461      462 
#> 163.3267 164.1082 167.9709 227.7443 166.8774 172.6750 167.4271 163.6517 
#>      463      464      465      466      467      468      469      470 
#> 163.6517 221.7342 165.2842 165.8723 165.8723 237.1116 180.4082 170.1703 
#>      471      472      473      474      475      476      477      478 
#> 166.4603 166.8514 173.0634 165.8723 176.9029 164.1777 221.1461 167.6226 
#>      479      480      481      482      484      485      486      487 
#> 222.1695 227.7962 166.4603 167.1048 164.1647 165.2842 163.5202 176.5377 
#>      488      489      490      491      492      493      494      495 
#> 169.5414 162.9322 170.3230 167.1873 166.1353 162.9322 164.1082 163.8990 
#>      496      497      498      499      500      501      502      503 
#> 166.4603 124.9258 224.1373 168.5589 220.5581 168.9377 221.1461 167.0456 
#>      504      505      506      507      508      509      510      511 
#> 171.8336 166.3076 173.0661 176.6111 170.9110 176.6676 174.7117 220.8211 
#>      512      514      515      516      517      518      519      520 
#> 166.3383 183.0007 174.3644 239.1277 175.4488 173.5976 178.1491 167.5401 
#>      521      522      523      524      525      526      527      528 
#> 173.7763 229.7818 221.1461 230.2263 177.0102 177.5702 164.6962 221.7342 
#>      529      530      531      532      533      534      535      536 
#> 221.1461 221.1461 164.1082 220.5581 162.9322 221.7342 164.6962 231.4023 
#>      537      538      539      540      541      542      543      544 
#> 161.7561 172.0870 165.8723 220.8804 164.6962 165.8723 171.0425 173.6587 
#>      545      546      547      548      549      550      551      552 
#> 169.7960 172.5789 167.6363 172.1910 174.7737 166.4603 165.8723 220.5581 
#>      553      554      555      556      557      558      559      560 
#> 162.4007 167.6363 168.2809 167.6363 169.7349 224.6736 220.5581 167.6363 
#>      561      562      563      564      565      566      567      568 
#> 164.6962 166.0038 179.4031 162.3441 168.8819 165.4157 176.1380 170.3230 
#>      569      570      571      572      573      574      575      576 
#> 164.5435 172.6003 178.0112 163.3675 174.1485 169.4692 169.2600 172.6750 
#>      577      578      579      580      581      582      583      584 
#> 164.6962 178.4900 174.8042 222.3222 165.4157 220.5581 165.9288 165.3408 
#>      585      586      587      588      589      590      591      592 
#> 165.2842 163.5202 167.9709 164.2472 164.6962 164.4305 162.9322 178.7371 
#>      593      594      595      596      597      599      600      601 
#> 165.8723 167.0483 176.4065 167.6363 178.8976 170.1183 161.7561 238.6222 
#>      602      603      604      605      606      607      608      609 
#> 162.9322 165.2842 172.0168 164.6962 165.2842 164.1082 173.1883 171.5685 
#>      610      611      612      613      614      615      616      617 
#> 223.0770 181.1671 171.4990 227.0035 186.2884 230.1086 167.3829 233.3054 
#>      618      619      620      621      622      623      624      625 
#> 162.9322 166.1945 220.5581 171.8336 220.5581 169.7349 180.0736 109.4223 
#>      626      627      628      630      631      632      633      634 
#> 186.7939 172.5483 225.0088 168.3497 233.4490 162.9322 166.8521 161.7561 
#>      635      636      637      638      639      640      641      642 
#> 173.6984 167.6363 225.0914 173.4021 168.1282 164.6962 168.6031 169.7349 
#>      643      644      645      646      647      648      649      650 
#> 172.1565 167.3829 167.3829 170.9110 167.6363 161.7561 168.4627 229.4595 
#>      651      652      653      654      655      656      657      659 
#> 123.5412 171.6503 173.5976 169.7349 230.6208 164.6962 162.9322 167.3829 
#>      660      661      662      663      664      665      666      667 
#> 229.2836 163.5202 182.7603 173.2726 221.7342 164.9592 167.0483 161.7561 
#>      669      670      671      672      673      674      675      676 
#> 170.0176 171.7114 221.9972 225.6579 162.9322 165.8723 220.5581 221.7342 
#>      677      678      679      680      681      682      683      684 
#> 171.3055 173.6587 167.6363 112.1915 233.4813 169.8357 172.5223 167.6363 
#>      685      686      687      688      689      690      691      692 
#> 169.1912 221.5249 112.6970 169.1469 168.2243 178.2270 117.1477 108.8343 
#>      693      694      695      696      697      698      699      700 
#> 167.7494 174.5690 170.0176 163.2544 172.7446 168.5589 164.1082 162.9322 
#>      701      702      703      704      705      706      707      708 
#> 167.9709 162.3441 167.9709 164.1082 167.3829 174.2864 164.1082 163.5202 
#>      709      710      711      712      713      714      715      716 
#> 167.6363 165.2842 164.8277 230.1301 222.1130 166.4603 163.8370 165.8723 
#>      717      718      719      720      722      723      724      725 
#> 167.0483 164.7527 171.0398 167.1048 174.3429 163.5202 221.1461 165.8723 
#>      726      727      728      729      730      731      732      733 
#> 178.4780 167.7678 167.3829 223.4020 166.5918 175.8750 166.7949 162.3441 
#>      734      735      736      737      738      740      741      742 
#> 167.7617 170.9110 169.1469 172.0870 165.2842 226.1849 168.9227 177.7215 
#>      743      744      745      746      747      748      749      750 
#> 220.5581 235.5002 177.5428 165.2788 162.9322 167.9709 167.6363 168.5070 
#>      751      752      753      754      755      756      757      758 
#> 165.8723 113.8730 168.9423 164.6962 174.7737 229.4595 165.6065 176.1804 
#>      759      760      761      762      763      764      765      766 
#> 167.9709 167.9839 232.3516 164.1082 169.1469 233.2183 167.9709 165.8723 
#>      767      769      770      772      773      774      775      776 
#> 168.5589 173.8319 222.9102 164.6962 167.5096 167.3829 221.1461 175.4442 
#>      777      778      779      780      781      782      783      784 
#> 164.5027 166.4603 165.3408 164.6962 168.2243 171.2112 224.9569 227.4869 
#>      785      786      787      788      789      790      791      792 
#> 165.2842 169.1469 182.3432 162.3441 167.1048 168.2243 223.7346 167.7678 
#>      793      794      795      796      797      798      799      800 
#> 165.2842 177.1257 166.1353 167.0483 221.7342 221.1461 173.0096 163.8990 
#>      801      802      803      804      805      806      807      808 
#> 173.2957 220.8211 168.0716 238.1486 166.4596 163.7832 164.1082 168.2243 
#>      809      810      811      812      813      814      815      816 
#> 178.8151 165.2842 174.8088 168.2243 220.5581 164.6962 174.3644 162.9322 
#>      817      818      819      820      821      822      823      824 
#> 165.7353 164.6962 174.3644 167.3829 179.4031 163.6517 171.4566 168.2243 
#>      825      826      827      828      829      830      831      832 
#> 185.7004 164.1082 164.8065 225.2616 165.6065 176.5455 222.3222 167.9709 
#>      833      834      835      836      837      838      839      840 
#> 172.6750 176.1804 171.4990 167.1737 166.1353 166.2068 165.8723 167.6363 
#>      841      842      843      844      845      846      847      848 
#> 165.2842 165.4157 168.2243 221.1461 166.4603 173.3747 165.8723 175.0738 
#>      849      850      851      852      853      854      855      856 
#> 221.1461 169.2240 170.6056 169.6082 177.1335 167.6363 165.2842 170.2915 
#>      857      858      859      860      861      862      863      864 
#> 220.5581 173.0707 221.7342 168.2243 169.1292 165.9288 166.1353 177.7215 
#>      865      866      867      868      869      870      871      872 
#> 225.7541 220.5581 170.9110 168.8689 162.4756 163.5202 183.9233 164.6962 
#>      873      874      875      876      877      878      879      880 
#> 164.6962 165.2842 164.7527 164.1082 163.5202 164.1082 169.6089 166.5141 
#>      881      882      883      885      886      887      888      889 
#> 163.5202 222.3222 221.8732 164.1082 230.7051 165.8723 172.0870 164.4305 
#>      890      891      892      893      894      895      896      897 
#> 170.3230 163.5202 169.8220 181.9598 161.7561 167.3706 180.0736 171.9908 
#>      898      899      900      901      904      906      907      908 
#> 182.7603 166.4603 161.7561 187.6600 179.4161 170.9110 164.2397 166.4603 
#>      909      910      911      912      913      914      915      916 
#> 185.1123 171.4990 165.2842 165.9853 174.3812 220.5581 165.2842 166.0576 
#>      917      918      919      920      921      922      923      924 
#> 229.3374 171.1195 168.2243 164.1082 165.1315 165.2842 164.1082 166.2068 
#>      925      926      927      928      929      930      932      933 
#> 171.6305 163.5202 163.5202 173.1013 169.1469 165.8723 174.7737 166.4603 
#>      934      935      936      937      938      939      940      941 
#> 169.4774 164.6962 166.4603 171.4243 163.5202 167.0483 178.8151 170.9110 
#>      942      943      944      945      946      947      948      949 
#> 176.5377 168.6414 166.4603 166.0113 168.5589 170.3230 225.0088 163.5202 
#>      951      952      953      954      955      956      957      958 
#> 170.7325 168.2243 165.2842 225.0914 183.9363 164.1082 168.4254 163.7832 
#>      959      960      961      962      963      964      965      966 
#> 166.4603 162.9322 170.8025 180.9962 164.6962 169.7349 167.3829 164.1082 
#>      967      968      969      970      971      972      973      974 
#> 220.5581 173.7209 231.2236 168.8689 226.1849 175.3617 165.4157 161.7561 
#>      975      976      977      978      979      980      981      982 
#> 169.2034 180.3744 172.8065 164.2370 221.7342 177.0601 164.6962 169.7349 
#>      983      984      985      986      987      988      989      990 
#> 168.5589 167.6363 166.3076 164.3167 169.1469 165.2788 165.0185 170.9110 
#>      991      992      993      994      995      996      997      998 
#> 165.6760 167.6363 167.6363 223.1187 171.1220 170.3268 166.4603 230.6356 
#>      999     1000     1001     1002     1003     1004     1005     1006 
#> 168.2303 221.7907 172.6750 169.1469 165.7945 165.2842 175.9360 161.7561 
#>     1007     1008     1009     1010     1011     1012     1013     1014 
#> 167.6363 166.4603 165.8723 166.9521 166.4603 162.9887 166.5918 168.6031 
#>     1015     1016     1017     1018     1019     1020     1021     1022 
#> 171.8336 165.2842 175.9497 166.4603 175.3617 164.7657 165.8723 228.9021 
#>     1023     1024     1026     1027     1028     1029     1030     1031 
#> 164.1082 182.1722 167.6363 168.0124 166.2048 169.4079 165.9233 171.2536 
#>     1032     1033     1034     1035     1036     1037     1038     1039 
#> 164.7527 167.0483 222.9102 163.9555 166.4603 170.9935 164.5027 167.3113 
#>     1040     1041     1042     1043     1044     1045     1046     1047 
#> 167.6363 166.0603 165.0185 225.0959 163.2544 169.3042 229.4595 163.5202 
#>     1048     1049     1050     1051     1052     1053     1054     1055 
#> 166.4603 178.2270 163.5202 164.8815 229.4595 226.7729 180.0171 171.4990 
#>     1056     1057     1058     1060     1061     1062     1063     1064 
#> 163.5202 175.6099 162.9322 163.5202 175.3617 170.3230 162.3441 162.9322 
#>     1065     1066     1067     1068     1069     1070     1071     1072 
#> 174.1856 167.6363 168.2243 174.1856 229.4595 164.6962 220.5581 220.5581 
#>     1073     1074     1075     1076     1077     1078     1079     1080 
#> 180.5609 166.4603 164.1082 162.9322 182.5647 173.9673 170.4802 169.7349 
#>     1081     1082     1083     1084     1085     1086     1087     1088 
#> 220.5581 168.8942 164.1082 165.6065 168.4874 163.5202 167.0483 164.6962 
#>     1089     1090     1091     1093     1094     1095     1096     1097 
#> 166.5168 164.6962 168.5466 175.8750 168.5635 172.0870 223.4982 166.2068 
#>     1098     1099     1100     1101     1102     1103     1104     1105 
#> 166.4603 222.9102 165.8723 174.2467 162.3441 162.9322 169.1912 167.0483 
#>     1106     1107     1108     1109     1110     1111     1112     1113 
#> 168.0320 166.7928 168.2100 167.3829 164.6962 177.5130 178.9632 167.0483 
#>     1114     1115     1116     1117     1118     1119     1120     1121 
#> 167.9709 163.5202 171.9343 225.2718 225.0088 221.7907 167.0483 167.1048 
#>     1122     1123     1124     1125     1126     1128     1129     1131 
#> 118.0051 183.3483 169.6089 170.3230 169.1469 162.9322 177.1868 167.6363 
#>     1132     1133     1134     1135     1136     1137     1138     1139 
#> 226.1849 169.7349 221.2027 166.4603 169.2477 164.6962 170.9675 165.2842 
#>     1140     1141     1142     1143     1144     1145     1146     1147 
#> 165.2222 165.5473 177.1855 169.8120 221.1461 162.0191 221.7342 174.5508 
#>     1148     1149     1150     1151     1152     1153     1154     1155 
#> 221.1434 232.0117 162.9322 166.7826 168.5589 166.4603 167.9709 166.4603 
#>     1156     1157     1158     1159     1160     1161     1162     1163 
#> 163.5202 173.0096 164.6962 168.2809 164.1082 171.4881 174.3172 165.8723 
#>     1164     1165     1166     1167     1168     1169     1170     1171 
#> 167.7466 170.5536 224.5033 168.2243 167.6363 161.7561 220.5581 165.8723 
#>     1172     1173     1174     1175     1176     1177     1178     1179 
#> 172.3131 165.8723 171.7620 163.5202 170.3230 167.9884 168.8812 164.6962 
#>     1180     1181     1182     1183     1184     1185     1186     1187 
#> 169.7915 167.6363 164.1082 167.3829 169.6675 175.2154 172.2956 222.3222 
#>     1188     1189     1190     1191     1192     1193     1194     1195 
#> 165.2842 232.9915 167.6363 168.2243 167.3829 168.2597 164.1647 163.5767 
#>     1196     1197     1198     1199     1200     1201     1202     1203 
#> 171.4990 167.6363 172.4111 167.6363 167.9709 167.0483 222.3222 179.9911 
#>     1204     1205     1206     1207     1208     1209     1210     1211 
#> 165.8723 168.6414 163.5174 220.8804 181.8867 165.8723 169.1469 164.6962 
#>     1212     1213     1214     1215     1216     1217     1218     1219 
#> 172.2723 225.0088 163.5202 164.3167 165.8723 162.9322 168.0716 225.0088 
#>     1220     1222     1223     1224     1225     1226     1227     1228 
#> 220.5581 166.3076 177.1257 163.5202 162.9322 168.2597 174.7737 163.5202 
#>     1229     1230     1231     1232     1233     1234     1236     1237 
#> 169.0202 231.1226 164.1082 165.8716 168.8689 181.5842 167.6363 170.4406 
#>     1238     1239     1240     1241     1242     1243     1244     1245 
#> 168.4874 185.6178 225.9914 169.7349 167.7678 163.5202 167.0483 174.9524 
#>     1246     1247     1248     1249     1250     1251     1252     1253 
#> 166.5298 179.4161 167.3829 169.8922 165.8103 161.7561 178.2270 225.0699 
#>     1254     1255     1256     1257     1258     1259     1260     1261 
#> 162.4007 228.9540 168.5589 222.1695 172.6003 169.1469 223.4982 165.2842 
#>     1262     1263     1264     1265     1266     1267     1268     1269 
#> 167.0483 173.1662 225.3311 164.6962 169.1469 234.1715 185.7004 170.3230 
#>     1270     1271     1272     1273     1274     1275     1276     1277 
#> 165.2993 164.6962 222.1825 167.0483 225.0088 172.1435 169.1469 162.9322 
#>     1278     1279     1280     1281     1282     1284     1285     1286 
#> 166.1945 167.0428 170.3230 173.0584 167.6363 225.0088 166.8548 168.6414 
#>     1288     1289     1290     1291     1292     1293     1294     1295 
#> 165.8723 167.6363 165.8723 165.8723 164.1082 176.7684 169.1469 175.4738 
#>     1296     1297     1298     1299     1300     1301     1302     1303 
#> 229.4595 168.9015 162.3441 163.5202 221.2027 167.0483 177.8029 172.5438 
#>     1304     1305     1306     1307     1308     1309     1310     1311 
#> 162.9322 165.9390 164.6962 167.4271 163.9147 164.8277 165.8723 168.5589 
#>     1312     1313     1314     1315     1316     1317     1318     1319 
#> 221.7342 179.5421 171.4990 221.7342 170.6575 167.3309 168.2243 162.9322 
#>     1320     1321     1322     1323     1324     1325     1326     1327 
#> 168.2781 168.5589 166.6483 183.8538 169.5857 167.0483 171.9726 109.4223 
#>     1328     1329     1330     1331     1333     1334     1335     1336 
#> 166.8391 163.5202 220.5581 221.7907 168.2243 165.8723 163.5202 166.1353 
#>     1337     1338     1339     1340     1341     1342     1343     1344 
#> 107.6583 165.2842 161.7561 170.9805 165.2842 170.8148 174.9524 165.8723 
#>     1345     1346     1347     1348     1349     1350     1351     1352 
#> 171.7666 167.0483 163.5202 164.1082 164.0250 172.4216 110.9375 177.3045 
#>     1353     1354     1355     1356     1357     1358     1359     1360 
#> 165.4233 175.3617 165.2842 164.0870 172.0123 162.9322 164.1082 166.4603 
#>     1361     1362     1363     1364     1365     1366     1367     1368 
#> 165.2222 226.6202 168.2216 165.2842 111.6782 169.4099 221.9195 162.9322 
#>     1369     1370     1371     1372     1373     1374     1375     1376 
#> 179.0236 168.8477 172.9057 169.5822 163.6592 221.8629 170.3230 168.8552 
#>     1377     1378     1379     1380     1381     1382     1383     1385 
#> 167.4271 162.0191 167.6363 224.6743 185.7004 176.5455 175.3923 177.6759 
#>     1386     1388     1389     1390     1391     1392     1393     1394 
#> 235.2650 166.4603 167.0483 169.7915 163.7832 164.3712 170.7583 163.5202 
#>     1395     1396     1397     1398     1399     1400     1401     1402 
#> 162.3441 170.7350 162.9322 176.7164 170.4463 169.5303 168.5466 167.6363 
#>     1403     1404     1405     1406     1407     1408     1409     1410 
#> 168.5589 163.5202 171.4990 165.2842 167.0483 174.2032 167.9709 167.6363 
#>     1411     1413     1414     1415     1416     1417     1418     1419 
#> 178.5753 164.7527 166.9521 166.4603 220.5581 163.2544 165.3613 184.5243 
#>     1420     1421     1422     1423     1424     1425     1426     1427 
#> 229.4595 165.5669 167.6363 164.2397 176.2629 112.6970 168.4874 164.1082 
#>     1428     1429     1430     1431     1432     1433     1434     1435 
#> 184.5243 167.0483 232.5099 223.5547 164.4305 167.6226 171.8687 164.1082 
#>     1436     1437     1438     1439     1440     1441     1442     1443 
#> 164.1082 168.2243 166.4603 168.5589 221.7342 167.6363 166.7948 162.9322 
#>     1444     1445     1446     1447     1448     1449     1450     1451 
#> 180.4829 166.3233 164.1082 162.9322 164.1082 174.7737 179.8202 225.0088 
#>     1452     1453     1454     1455     1456     1457     1458     1459 
#> 165.3408 162.9322 220.5581 222.1287 162.3441 172.1046 165.2842 166.7948 
#>     1460     1461     1462     1463     1464     1465     1466     1467 
#> 173.7763 169.7349 172.0446 167.6363 165.2842 165.2630 166.9521 167.0483 
#>     1468     1469     1470     1471     1472     1474     1475     1476 
#> 166.4603 164.1082 166.2068 166.2668 167.6363 164.1082 177.0510 167.6336 
#>     1477     1478     1479     1480     1481     1482     1483     1484 
#> 163.5202 166.7948 172.6750 236.9413 167.9709 173.5976 168.2809 175.6997 
#>     1485     1486     1487     1488     1489     1490     1491     1492 
#> 222.3222 163.8990 172.0870 162.9322 230.7746 164.6962 171.6120 165.2842 
#>     1493     1494     1495     1496     1497     1498     1499     1500 
#> 162.9322 173.0584 173.0096 162.9322 221.7342 175.9497 175.7970 164.6962 
#>     1501     1502     1503     1504     1505     1506     1507     1508 
#> 166.9521 172.6750 221.7342 162.9322 172.8038 164.1082 164.1082 165.3538 
#>     1509     1510     1511     1512     1513     1514     1515     1516 
#> 167.5401 166.6688 168.0151 166.4603 222.9102 169.8045 182.7377 170.9110 
#>     1517     1518     1519     1520     1521     1522     1523     1524 
#> 165.2630 222.9102 167.6363 175.2154 169.0849 171.2455 165.4103 169.9202 
#>     1525     1526     1527     1528     1529     1530     1531     1532 
#> 170.6056 221.8472 163.5202 220.5581 223.9901 165.0158 169.7349 166.4603 
#>     1533     1534     1535     1536     1537     1538     1539     1540 
#> 166.4603 168.2243 165.9513 175.8877 170.9675 166.8548 170.2145 226.1849 
#>     1541     1542     1543     1544     1545     1546     1547     1549 
#> 177.0510 162.3441 167.0483 235.8530 164.3712 175.0219 177.8268 167.6363 
#>     1550     1551     1552     1553     1554     1555     1556     1557 
#> 166.7948 164.1082 220.5581 167.8993 169.1469 166.5993 239.2102 167.2363 
#>     1558     1559     1560     1561     1562     1563     1564     1565 
#> 163.2544 123.9357 163.5202 167.0483 165.9288 125.5138 166.3076 166.4603 
#>     1566     1567     1568     1569     1570     1571     1572     1573 
#> 164.4305 168.2243 167.6363 167.9709 162.9322 181.2497 165.2842 220.8804 
#>     1574     1575     1576     1577     1578     1579     1580     1581 
#> 164.1082 170.7018 168.2243 168.6414 165.2842 169.4542 177.3888 169.2470 
#>     1582     1584     1585     1586     1587     1588     1589     1590 
#> 221.7342 167.0483 177.8925 108.8343 170.2268 167.6363 165.8723 165.8723 
#>     1591     1592     1593     1594     1596     1597     1598     1599 
#> 170.1703 169.5822 169.9656 171.2455 165.4157 163.5202 174.7374 175.9497 
#>     1600     1601     1602     1603     1604     1605     1606     1607 
#> 165.8723 181.7552 171.4990 181.7552 168.5589 164.6962 178.2270 165.8723 
#>     1608     1609     1610     1611     1613     1614     1615     1616 
#> 167.3829 221.1461 167.3309 165.1315 166.8548 222.3222 164.6962 170.9675 
#>     1617     1618     1619     1620     1621     1622     1623     1624 
#> 164.8277 167.6363 169.6388 176.9029 163.3212 166.4603 166.4603 167.0483 
#>     1625     1626     1627     1628     1629     1630     1631     1632 
#> 167.6459 167.0483 177.8528 177.5130 163.5897 165.8723 165.8723 170.9110 
#>     1633     1634     1635     1636     1637     1638     1639     1640 
#> 164.2472 186.2059 171.3846 129.3069 166.2068 165.8723 166.7826 168.2243 
#>     1641     1642     1643     1644     1645     1646     1647     1648 
#> 162.9322 165.2842 163.5202 221.1461 167.6363 165.7196 166.6558 168.2932 
#>     1649     1650     1651     1652     1653     1654     1655     1656 
#> 167.0483 169.7349 167.6363 161.7561 168.5589 162.9887 166.4603 164.1082 
#>     1657     1658     1659     1660     1661     1662     1663     1664 
#> 163.5202 162.4007 170.8946 180.7838 174.7737 226.2979 220.8804 176.0277 
#>     1665     1666     1667     1668     1669     1670     1671     1672 
#> 165.2842 175.8535 165.1473 164.3712 170.3795 163.5202 165.2842 166.4603 
#>     1673     1674     1675     1677     1678     1679     1680     1681 
#> 225.0088 162.9322 167.6363 162.4137 166.2068 170.3230 109.4223 169.9980 
#>     1682     1683     1684     1685     1686     1687     1688     1689 
#> 232.4691 220.5581 175.9497 165.4928 166.4603 163.5202 164.1082 116.5948 
#>     1690     1691     1692     1693     1694     1695     1696     1697 
#> 167.6363 231.5459 184.4418 225.6579 229.2020 162.9322 167.9586 162.9322 
#>     1698     1699     1700     1701     1702     1703     1704     1705 
#> 167.0483 222.5852 162.9322 221.7342 161.7561 166.4603 164.6962 170.9110 
#>     1706     1707     1708     1709     1710     1711     1713     1714 
#> 220.5581 171.8336 164.1082 171.4028 173.5976 166.8514 170.9110 162.9322 
#>     1715     1716     1717     1718     1719     1720     1721     1722 
#> 226.5071 172.0870 222.3222 167.8781 174.9081 162.9322 172.6750 165.6630 
#>     1723     1724     1725     1726     1727     1728     1729     1730 
#> 174.5708 172.6003 226.2674 173.0096 167.0483 221.1461 173.5976 163.5202 
#>     1731     1732     1734     1735     1736     1737     1738     1739 
#> 221.2157 167.6363 163.5202 162.9322 165.2842 168.5589 167.5144 175.5404 
#>     1740     1741     1742     1743     1744     1745     1746     1747 
#> 164.1082 163.5202 162.9322 165.2842 171.8336 222.3222 164.0462 168.9534 
#>     1748     1749     1750     1751     1752     1753     1754     1755 
#> 164.6962 164.4305 162.9322 166.5168 166.2511 221.2157 124.9258 171.4028 
#>     1756     1757     1758     1759     1760     1761     1762     1763 
#> 165.0185 166.4603 163.5202 164.1082 171.4990 163.5202 166.1353 164.6962 
#>     1764     1765     1766     1767     1768     1769     1770     1772 
#> 225.7849 164.1082 165.2842 163.8425 164.1082 164.9592 237.2376 166.4603 
#>     1773     1774     1775     1776     1777     1778     1779     1780 
#> 165.0178 166.5373 167.9089 172.4216 170.4620 164.1082 171.4853 168.2243 
#>     1782     1783     1784     1785     1786     1787     1788     1789 
#> 167.3829 162.0784 163.5202 221.1461 225.2623 168.2243 169.7349 164.2397 
#>     1790     1791     1792     1793     1794     1795     1796     1797 
#> 169.4295 162.9322 167.3706 177.4435 174.9478 170.1721 169.0127 167.3829 
#>     1798     1799     1800     1801     1802     1803     1804     1805 
#> 170.3230 173.5976 167.0911 167.0483 167.9709 175.8729 165.2842 162.9322 
#>     1806     1807     1808     1809     1810     1811     1812     1813 
#> 171.4990 225.9756 222.4455 166.5298 168.5589 162.3441 172.6750 175.5404 
#>     1814     1815     1816     1817     1818     1819     1820     1821 
#> 173.4449 185.2384 171.4028 164.6962 165.8695 162.9322 174.4430 168.2243 
#>     1822     1823     1824     1825     1827     1828     1829     1830 
#> 178.8151 167.6363 165.8723 163.5202 163.5202 162.3441 166.8956 171.8336 
#>     1831     1832     1833     1834     1836     1837     1838     1839 
#> 167.6363 167.0483 165.3408 232.4534 171.7855 170.7583 168.2932 168.2243 
#>     1840     1841     1842     1843     1844     1845     1846     1847 
#> 181.6362 168.0151 220.5581 224.9569 165.0185 162.9322 164.5722 126.3103 
#>     1848     1849     1851     1852     1853     1854     1855     1856 
#> 166.5902 166.2068 170.9110 168.2243 163.6517 169.0542 220.5581 222.1695 
#>     1857     1858     1859     1860     1861     1862     1863     1864 
#> 222.1695 166.4603 165.2842 165.8723 168.5589 168.8219 172.5789 221.7342 
#>     1865     1866     1867     1868     1869     1870     1871     1872 
#> 162.9322 220.5581 165.2842 167.1048 173.1357 164.1082 164.6962 167.0911 
#>     1873     1875     1876     1877     1878     1879     1880     1881 
#> 165.2842 167.7058 222.3222 225.5968 223.4362 169.4295 221.1461 176.2674 
#>     1882     1883     1884     1885     1886     1887     1888     1889 
#> 230.0476 224.6743 174.6775 174.3644 175.9497 168.8096 173.0096 176.5377 
#>     1890     1891     1892     1893     1894     1895     1896     1897 
#> 232.1950 169.2294 168.2107 229.4595 222.1130 222.3787 166.3983 167.6363 
#>     1898     1899     1900     1901     1902     1903     1904     1905 
#> 167.1798 220.5581 220.5581 172.6750 164.8277 167.6363 172.6750 163.8425 
#>     1906     1907     1908     1909     1910     1911     1912     1913 
#> 163.5202 168.2243 162.9322 169.5264 170.9110 162.3441 173.5976 169.7349 
#>     1914     1915     1916     1917     1918     1919     1920     1921 
#> 177.8925 164.1082 174.1856 165.6788 167.9709 167.1178 164.7527 220.5581 
#>     1922     1923     1924     1925     1926     1927     1928     1929 
#> 167.9688 178.2530 165.5623 176.1165 162.3441 170.9110 176.8555 165.1500 
#>     1930     1931     1932     1933     1934     1935     1936     1937 
#> 167.6363 222.9102 163.5202 168.2243 172.1435 165.4853 231.2819 163.5202 
#>     1938     1939     1940     1941     1942     1943     1944     1945 
#> 167.9709 167.6363 167.0483 167.9189 168.6719 220.8804 168.8689 167.9709 
#>     1946     1947     1948     1949     1950     1951     1952     1953 
#> 119.9935 178.8151 165.4157 222.9102 224.6743 177.3045 166.4603 221.4657 
#>     1954     1955     1956     1957     1959     1960     1961     1962 
#> 166.8493 220.5581 167.7678 167.6363 164.1082 164.6962 164.3712 167.7651 
#>     1963     1964     1965     1966     1967     1968     1969     1970 
#> 165.8723 165.0750 165.2842 167.5531 165.2842 170.8148 172.4781 164.1082 
#>     1971     1972     1973     1974     1975     1976     1977     1978 
#> 167.2438 166.2068 170.1295 167.1736 177.0510 220.5581 170.1580 168.2243 
#>     1979     1980     1981     1982     1983     1984     1985     1986 
#> 222.3787 173.3747 165.2842 166.0733 173.5976 164.6962 172.4658 164.1082 
#>     1987     1988     1989     1990     1991     1992     1993     1994 
#> 164.5027 176.9250 229.4595 164.1082 173.0096 171.2898 177.8261 164.6962 
#>     1995     1996     1997     1998     1999     2000 
#> 163.7212 223.4982 224.7438 162.6664 166.7826 169.1469